Load previous data entry in form controller
