+++ frontend/app/component/form/data_entry/differences/DifferencesForm.test.tsx
@@ -177,7 +177,7 @@ describe("Test DifferencesForm", () => {
         ...expectedRequest.data.differences_counts,
       });
 
-      const submitButton = screen.getByRole("button", { name: "Volgende" });
+      const submitButton = await screen.findByRole("button", { name: "Volgende" });
       await user.click(submitButton);
 
       expect(spy).toHaveBeenCalled();
@@ -201,7 +201,7 @@ describe("Test DifferencesForm", () => {
 
       renderForm({ recounted: false });
 
-      const submitButton = screen.getByRole("button", { name: "Volgende" });
+      const submitButton = await screen.findByRole("button", { name: "Volgende" });
       await user.click(submitButton);
 
       const feedbackMessage =
@@ -235,7 +235,7 @@ describe("Test DifferencesForm", () => {
 
       renderForm({ recounted: true });
 
-      const submitButton = screen.getByRole("button", { name: "Volgende" });
+      const submitButton = await screen.findByRole("button", { name: "Volgende" });
       await user.click(submitButton);
 
       const feedbackMessage =
@@ -269,7 +269,7 @@ describe("Test DifferencesForm", () => {
 
       renderForm({ recounted: false });
 
-      const submitButton = screen.getByRole("button", { name: "Volgende" });
+      const submitButton = await screen.findByRole("button", { name: "Volgende" });
       await user.click(submitButton);
 
       const feedbackMessage =
@@ -303,7 +303,7 @@ describe("Test DifferencesForm", () => {
 
       renderForm({ recounted: true });
 
-      const submitButton = screen.getByRole("button", { name: "Volgende" });
+      const submitButton = await screen.findByRole("button", { name: "Volgende" });
       await user.click(submitButton);
 
       const feedbackMessage =
@@ -349,7 +349,7 @@ describe("Test DifferencesForm", () => {
 
       renderForm({ recounted: false });
 
-      const submitButton = screen.getByRole("button", { name: "Volgende" });
+      const submitButton = await screen.findByRole("button", { name: "Volgende" });
       await user.click(submitButton);
 
       const feedbackMessage =
@@ -397,7 +397,7 @@ describe("Test DifferencesForm", () => {
 
       renderForm({ recounted: false });
 
-      const submitButton = screen.getByRole("button", { name: "Volgende" });
+      const submitButton = await screen.findByRole("button", { name: "Volgende" });
       await user.click(submitButton);
 
       const feedbackMessage =
@@ -464,7 +464,7 @@ describe("Test DifferencesForm", () => {
 
       renderForm({ recounted: false });
 
-      const submitButton = screen.getByRole("button", { name: "Volgende" });
+      const submitButton = await screen.findByRole("button", { name: "Volgende" });
       await user.click(submitButton);
 
       const feedbackMessage =
@@ -510,7 +510,7 @@ describe("Test DifferencesForm", () => {
 
       renderForm({ recounted: false });
 
-      const submitButton = screen.getByRole("button", { name: "Volgende" });
+      const submitButton = await screen.findByRole("button", { name: "Volgende" });
       await user.click(submitButton);
 
       const feedbackMessage =
